feat!: upgrade dependencies
