--- circle.yml
@@ -1,49 +1,0 @@
-version: 2
-jobs:
-  build:
-    working_directory: ~/rereselect
-    environment:
-      CIRCLE_ARTIFACTS: /tmp/circleci-artifacts
-      CIRCLE_TEST_REPORTS: /tmp/circleci-test-results
-    docker:
-      - image: circleci/node:10
-    steps:
-      - checkout
-      - run: mkdir -p $CIRCLE_ARTIFACTS $CIRCLE_TEST_REPORTS
-      - restore_cache:
-          keys:
-            - v1-dep-{{ .Branch }}-
-            - v1-dep-master-
-            - v1-dep-
-      - run: yarn
-      - save_cache:
-          key: v1-dep-{{ .Branch }}-{{ epoch }}
-          paths:
-            - node_modules
-            - ~/.cache/yarn
-      - run:
-          name: prepare
-          command: |
-            yarn prepare
-      - run:
-          name: test (compiled)
-          command: |
-            yarn test
-      - run:
-          name: test (source)
-          command: |
-            yarn test:cov
-      - run:
-          name: benchmark
-          command: |
-            yarn benchmark
-      - run:
-          name: codecov
-          command: |
-            curl -s https://codecov.io/bash | bash
-      - store_test_results:
-          path: /tmp/circleci-test-results
-      - store_artifacts:
-          path: /tmp/circleci-artifacts
-      - store_artifacts:
-          path: /tmp/circleci-test-results
