--- lib/index.test.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=index.test.d.ts.map
