--- tsconfig.es6.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "target": "es2015",
-    "outDir": "lib-es6"
-  }
-}
